requirements.txt
huggingface_hub
numpy
Pillow
tensorflow
tkinterdnd2
